Add QPACK dynamic table support (RFC 9204)
Introduce Encoder/Decoder, backed by a real DynamicTable and the
encoder-stream/decoder-stream instruction codecs, alongside the
existing static-table-only SimpleEncoder/SimpleDecoder (renamed from
Encoder/Decoder to make room). Encoder encodes in strictly
non-blocking mode; Decoder mirrors the peer's table and decodes any
compliant encoding, including dynamic-table and post-Base references.

Verified byte-for-byte against the RFC 9204 Appendix B.1 and B.2-B.4
worked examples, plus a full encoder/decoder round trip across
multiple field sections.

+++ src/org/bluezoo/gumdrop/http/qpack/PrefixedInteger.java
@@ -52,14 +52,14 @@ final class PrefixedInteger {
      * @param prefixBits the number of low bits of the first byte
      *                   available for the value, 1-8
      */
-    static void encode(ByteBuffer buf, int highBits, int value, int prefixBits) {
+    static void encode(ByteBuffer buf, int highBits, long value, int prefixBits) {
         int prefixMax = (1 << prefixBits) - 1;
         if (value < prefixMax) {
             buf.put((byte) (highBits | value));
             return;
         }
         buf.put((byte) (highBits | prefixMax));
-        int remaining = value - prefixMax;
+        long remaining = value - prefixMax;
         while (remaining >= 128) {
             buf.put((byte) ((remaining % 128) | 0x80));
             remaining /= 128;
